[MISC] move seqan3::views::single_pass_input to seqan3/utility/views/
