+++ include/seqan3/io/sam_file/detail/cigar.hpp
@@ -21,10 +21,10 @@
 #include <seqan3/alignment/detail/pairwise_alignment_concept.hpp>
 #include <seqan3/alphabet/cigar/cigar.hpp>
 #include <seqan3/core/debug_stream/detail/to_string.hpp>
-#include <seqan3/range/views/single_pass_input.hpp>
 #include <seqan3/range/views/take_until.hpp>
 #include <seqan3/utility/char_operations/predicate.hpp>
 #include <seqan3/utility/tuple/concept.hpp>
+#include <seqan3/utility/views/single_pass_input.hpp>
 #include <seqan3/utility/views/zip.hpp>
 
 namespace seqan3::detail
