+++ test/performance/range/views/view_drop_benchmark.cpp
@@ -15,7 +15,7 @@
 #include <benchmark/benchmark.h>
 
 #include <seqan3/range/views/drop.hpp>
-#include <seqan3/range/views/single_pass_input.hpp>
+#include <seqan3/utility/views/single_pass_input.hpp>
 
 // ============================================================================
 //  sequential_read
