+++ test/performance/range/views/view_take_until_benchmark.cpp
@@ -14,8 +14,8 @@
 #include <benchmark/benchmark.h>
 
 #include <seqan3/range/views/take_until.hpp>
-#include <seqan3/range/views/single_pass_input.hpp>
 #include <seqan3/utility/char_operations/predicate.hpp>
+#include <seqan3/utility/views/single_pass_input.hpp>
 
 // ============================================================================
 //  sequential_read
