+++ test/snippet/utility/views/single_pass_input.cpp
@@ -1,7 +1,7 @@
 #include <string>
 
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/range/views/single_pass_input.hpp>
+#include <seqan3/utility/views/single_pass_input.hpp>
 
 int main()
 {
