+++ test/unit/io/views/async_input_buffer_test.cpp
@@ -16,9 +16,9 @@
 #include <seqan3/core/detail/debug_stream_alphabet.hpp>
 #include <seqan3/io/views/async_input_buffer.hpp>
 #include <seqan3/range/concept.hpp>
-#include <seqan3/range/views/single_pass_input.hpp>
 #include <seqan3/range/views/take.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
+#include <seqan3/utility/views/single_pass_input.hpp>
 
 #include "../../range/iterator_test_template.hpp"
