+++ test/unit/range/views/CMakeLists.txt
@@ -22,6 +22,5 @@ seqan3_test(to_upper_test.cpp)
 seqan3_test(translate_join_test.cpp)
 seqan3_test(translate_test.cpp)
 seqan3_test(trim_quality_test.cpp)
-seqan3_test(single_pass_input_test.cpp)
 seqan3_test(get_test.cpp)
 seqan3_test(interleave_test.cpp)
