+++ test/unit/utility/views/CMakeLists.txt
@@ -3,4 +3,5 @@ add_subdirectories()
 seqan3_test (enforce_random_access_test.cpp)
 seqan3_test (repeat_n_test.cpp)
 seqan3_test (repeat_test.cpp)
+seqan3_test (single_pass_input_test.cpp)
 seqan3_test (zip_test.cpp)
