+++ test/unit/utility/views/single_pass_input_test.cpp
@@ -14,8 +14,8 @@
 #include <range/v3/view/slice.hpp>
 
 #include <seqan3/core/detail/persist_view.hpp>
-#include <seqan3/range/views/single_pass_input.hpp>
 #include <seqan3/test/expect_same_type.hpp>
+#include <seqan3/utility/views/single_pass_input.hpp>
 #include <seqan3/utility/views/zip.hpp>
 
 template <typename rng_type>
